feat: added my events page, lists all events related to the authenticated user
